Day-2/CrewAI/requirements.txt
crewai
crewai[tools]
python-dotenv
sentence-transformers